Extract graphic container out of BannerModal.
